[hdSt] Fix for dome light shading differences between GL and Vulkan.
The sinTheta term calculated during generation of the BRDF texture can be NaN
with Vulkan (rather than 0 with GL). This seems to be happening due to floating
point imprecision. When Xi.y values are close to 1 and roughness values are
close to 0, the cosTheta term is very close to 1 for both GL and Vulkan (such
that evaluating cosTheta == 1.0 returns true). However, cosTheta * cosTheta
will still equal 1 for GL, but will be > 1 for Vulkan (leading us to take the
sqrt of a negative value, resulting in the NaN). To fix this, we stop cosTheta
from being greater than 1.

(Internal change: 2369559)