Implement tombstoning of uncompilation rules when they refer to the same VdfInput.
This fixes the issue where Relationship().TargetedObjects() inputs would
be uncompiled on changes to previously targeted objects.

Exec_UncompilationRule has been refactored to consist of EsfEditReasons and
an Exec_UncompilationTarget. The target is a std::variant that can be one
of Exec_NodeUncompilationTarget or Exec_InputUncompilationTarget.

Exec_NodeUncompilationTargets identify a node to be uncompiled by its VdfId.
Once a node has been uncompiled, other rules for the same node are slated
for garbage collection, because their VdfIds are no longer valid.

Exec_InputUncompilationTargets identify an input to be uncompiled. This is
represented as a pointer to a shared "identity" object. The identity
describes the input to be uncompiled (as a nodeId/inputName pair), as well
as an "isValid" flag that becomes false when the input is uncompiled. The
first time an Exec_InputUncompilationTarget is uncompiled, the "isValid"
flag gets cleared, which prevents other rules from uncompiling the same
input in subsequent rounds of change processing.

(Internal change: 2369853)