pcp: Fix strength ordering change with simplified specializes
The new simplified specializes implementation in change 2369111
inadvertently changed the strength ordering of opinions in a
production asset using a tricky combination of specializes and
ancestral references and payloads. This was caused by changes
to the order in which certain tasks were evaluated, which hit
a known issue where multiple composition arcs to the same site
are handled with a "first one wins" policy instead of something
more principled.

This change introduces a "strongest origin" policy for cases
involving implied specializes. This is a targeted fix to
address this specific issue, but could expand to be more
general in the future. That's a much bigger change, though.

(Internal change: 2369883)