tf: Add a context manager & decorator that will catch raised Tf.ErrorExceptions, re-post the contained TfErrors back to the thread's error list, and stop propagating the Python exception.  This provides a way to sidestep PyQt/SIP's behavior of blocking raised exceptions from virtual function overrides so that TfErrors can continue propagating to callers.
Also add a `RaiseIfNotClean()` member on Tf.ErrorMark (Python-only) that will
raise contained errors as a Tf.ErrorException, if there are any.  This can be
useful in cases where Tf.Errors were generated during a call through a foreign
C++ binding: the caller can collect them with a Tf.ErrorMark and then call
RaiseIfNotClean() to propagate them as a Tf.ErrorException as usual.

Also a tiny fix for finding code locations when globals has no
'__name__', which turns out is the case in usdview's interpreter.

(Internal change: 2370143)