Change Exec_OutputKey::Identity to use the computation definition pointer instead of the computation name.
This is necessary to support dispatched computations, which will otherwise have
name collisions when two distinct computations with the same name are dispatched
onto the same prim.

It is currently safe to use raw pointers to computation definitions in this way
because computation definitions are immortal. Even if we evict them when we add
support for configuration, we think this is safe because we will need to evict
affected output key identities from exec data structures, or else they would key
stale cache entries.

(Internal change: 2370249)