[HdMtlx, HdSt-Mx] Fix error in MaterialXFilter
Storm needs to have all the primvars connected to the terminal
node in order for them to be caught in the material network. The
function that was cycling through the nodes was doing this
incorrectly. Fixing that highlighted a fallout of the v1.38 -> v1.39
node change issues, in that at this point we might be asking v1.39
for a v1.38 node, so this change exposes the HdMtlx function that
updates the nodeDef name.

(Internal change: 2370558)